refactor to authenticate user in controller instead of the service

+++ backend/src/main/java/com/example/toMeMail/service/MessageServiceImpl.java
@@ -5,8 +5,12 @@ import com.example.toMeMail.entity.User;
 import com.example.toMeMail.exception.UserNotFoundException;
 import com.example.toMeMail.repository.MessageRepository;
 import com.example.toMeMail.repository.UserRepository;
+import com.example.toMeMail.security.CustomUserDetails;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,32 +27,46 @@ public class MessageServiceImpl implements MessageService {
 
     @Override
     @Transactional
-    public Message createMessage(Message message, String username) {
+    public Message createMessage(Message message) {
 
-        User user = userRepository.findByUsername(username).orElseThrow(() -> new UserNotFoundException("User " + username + " not found"));
+        User user = userRepository.findByUsername(getAuthenticatedUsername()).orElseThrow(() -> new UserNotFoundException("User not found"));
         message.setUser(user);
         return messageRepository.save(message);
     }
 
     @Override
-    public List<Message> getMessagesByUsername(String username) {
-        return messageRepository.findByUserUsername(username);
+    public List<Message> getMessagesByUsername() {
+        return messageRepository.findByUserUsername(getAuthenticatedUsername());
     }
 
     @Override
-    public Message getMessageById(Long messageId, String username) {
-        Message message = messageRepository.findById(messageId).orElseThrow(() -> new UserNotFoundException("User with ID " + username + " not found"));
+    public Message getMessageById(Long messageId) {
+        Message message = messageRepository.findById(messageId).orElseThrow(() -> new UserNotFoundException("User not found"));
         /*if (message.getUser().getId() != userId) {
             throw new RuntimeException("No authorized to access this message!");
         }*/
+        // Check if the message belongs to the authenticated user
+        if (!message.getUser().getUsername().equals(getAuthenticatedUsername())) {
+            throw new AccessDeniedException("You are not authorized to access this message");
+        }
         return message;
     }
 
     @Override
     @Transactional
-    public void deleteMessage(Long messageId, String username) {
+    public void deleteMessage(Long messageId) {
         Message message = messageRepository.findById(messageId).orElseThrow(() -> new RuntimeException("Message not found"));
 
+        if (!message.getUser().getUsername().equals(getAuthenticatedUsername())) {
+            throw new AccessDeniedException("You are not authorized to access this message");
+        }
+
         messageRepository.delete(message);
+    }
+
+    private String getAuthenticatedUsername() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        return userDetails.getUsername();
     }
 }
